Upgrade importlib-metadata before installing twine
In Python 3.7, `make deps` logs the following message:

    Requirement already satisfied: importlib-metadata; python_version < "3.8"
    in /home/travis/virtualenv/python3.7.1/lib/python3.7/site-packages
    (from twine) (0.18)

The `importlib-metadata` is an indirect dependency. The above message
shows that `pip` does not install the latest version of
`importlib-metadata` because an old version of it is already present in
the virtual Python environment that comes with Travis CI. Due to this a
version conflict occurs with another indirect dependency which is then
logged later as follows:

    ERROR: keyring 22.0.1 has requirement importlib-metadata>=1;
    python_version < "3.8", but you'll have importlib-metadata 0.18
    which is incompatible.

As a result of this, later during `twine check dist/*` the following
error occurs:

    pkg_resources.ContextualVersionConflict: (importlib-metadata 0.18
    (/home/travis/virtualenv/python3.7.1/lib/python3.7/site-packages),

    Requirement.parse('importlib-metadata>=1; python_version < "3.8"'),
    {'keyring'})

    Makefile:74: recipe for target 'dist' failed

As a workaround, this change now upgrades `importlib-metadata` and then
installs the other dependencies.